Check A  Number Even Or Odd.
This Python program determines whether a number is even or odd. It asks the user to input a number, then checks if the number is divisible by 2. If the number is divisible by 2 (i.e., the remainder is 0), it is classified as even; otherwise, it is classified as odd. The program then displays a message indicating whether the entered number is even or odd.